Added the Authentication Token for CRUD Operations on Appliances

+++ lib/AuthService/auth_appliances.dart
@@ -81,7 +81,7 @@ class ApplianceService {
     }
 
     final url = Uri.parse(
-        'http://10.0.2.2:8080/getAllUsersAppliances/$userId/appliances');
+        '${ApiConfig.baseUrl}/getAllUsersAppliances/$userId/appliances');
 
     final response = await http.get(url);
 
@@ -109,37 +109,45 @@ class ApplianceService {
 
     final url = Uri.parse(
         '${ApiConfig.baseUrl}/updateApplianceOccurrences/$applianceId');
+    String? token = await getToken();
+    if (token != null) {
+      final response = await http.patch(
+        url,
+        headers: <String, String>{
+          'Content-Type': 'application/json; charset=UTF-8',
+          'Authorization': 'Bearer $token',
+        },
+        body: jsonEncode(updates),
+      );
 
-    final response = await http.patch(
-      url,
-      headers: <String, String>{
-        'Content-Type': 'application/json; charset=UTF-8',
-      },
-      body: jsonEncode(updates),
-    );
-
-    if (response.statusCode == 200) {
-      final responseBody = jsonDecode(response.body);
-      return responseBody; // Return response if needed
-    } else {
-      final responseBody = jsonDecode(response.body);
-      throw Exception('Failed to update appliance: ${responseBody['error']}');
+      if (response.statusCode == 200) {
+        final responseBody = jsonDecode(response.body);
+        return responseBody; // Return response if needed
+      } else {
+        final responseBody = jsonDecode(response.body);
+        throw Exception('Failed to update appliance: ${responseBody['error']}');
+      }
     }
   }
 
   // Static method to delete an appliance
   static Future<void> deleteAppliance(String applianceId) async {
     final url = Uri.parse('${ApiConfig.baseUrl}/deleteAppliance/$applianceId');
-    final response = await http.delete(
-      url,
-      headers: <String, String>{
-        'Content-Type': 'application/json; charset=UTF-8',
-      },
-    );
-    if (response.statusCode == 200) {
-    } else {
-      final responseBody = jsonDecode(response.body);
-      throw Exception('Failed to delete appliance: ${responseBody['message']}');
+    String? token = await getToken();
+    if (token != null) {
+      final response = await http.delete(
+        url,
+        headers: <String, String>{
+          'Content-Type': 'application/json; charset=UTF-8',
+          'Authorization': 'Bearer $token',
+        },
+      );
+      if (response.statusCode == 200) {
+      } else {
+        final responseBody = jsonDecode(response.body);
+        throw Exception(
+            'Failed to delete appliance: ${responseBody['message']}');
+      }
     }
   }
